{{cookiecutter.project_slug}}/requirements/testing.txt
-r default.txt

asynctest==0.13.0
pytest==5.2.2
pytest-aiohttp==0.3.0
pytest-asyncio==0.10.0
pytest-cov==2.8.1
